canBe series를 별도 함수로 분리합니다
