+++ src/index.ts
@@ -1,19 +1,12 @@
 export { assembleHangul } from './assemble';
-export { chosungIncludes } from './chosungIncludes';
+export { canBeChoseong, canBeJongseong, canBeJungseong } from './canBe';
 export { choseongIncludes } from './choseongIncludes';
+export { chosungIncludes } from './chosungIncludes';
 export { combineHangulCharacter, combineVowels, curriedCombineHangulCharacter } from './combineHangulCharacter';
 export { convertQwertyToHangul, convertQwertyToHangulAlphabet } from './convertQwertyToHangulAlphabet';
 export { disassemble, disassembleToGroups } from './disassemble';
 export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
+export { getChoseong } from './getChoseong';
 export { josa } from './josa';
 export { removeLastHangulCharacter } from './removeLastHangulCharacter';
-export { getChoseong } from './getChoseong';
-export {
-  canBeChosung,
-  canBeJongsung,
-  canBeJungsung,
-  hasBatchim,
-  hasProperty,
-  hasSingleBatchim,
-  hasValueInReadOnlyStringList,
-} from './utils';
+export { hasBatchim, hasProperty, hasSingleBatchim, hasValueInReadOnlyStringList } from './utils';
