+++ src/utils.ts
@@ -1,9 +1,7 @@
 import {
-  COMPLETE_HANGUL_START_CHARCODE,
   COMPLETE_HANGUL_END_CHARCODE,
-  HANGUL_CHARACTERS_BY_FIRST_INDEX,
+  COMPLETE_HANGUL_START_CHARCODE,
   HANGUL_CHARACTERS_BY_LAST_INDEX,
-  HANGUL_CHARACTERS_BY_MIDDLE_INDEX,
   NUMBER_OF_JONGSEONG,
 } from './constants';
 
@@ -67,139 +65,6 @@ export function hasSingleBatchim(str: string) {
 
   const batchimCode = (charCode - COMPLETE_HANGUL_START_CHARCODE) % NUMBER_OF_JONGSEONG;
   return HANGUL_CHARACTERS_BY_LAST_INDEX[batchimCode].length === 1;
-}
-
-/**
- * @name canBeChosung
- * @deprecated canBeChoseong을 사용해 주세요.
- * @description
- * 인자로 받은 문자가 초성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeChosung(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeChosung('ㄱ') // true
- * canBeChosung('ㅃ') // true
- * canBeChosung('ㄱㅅ') // false
- * canBeChosung('ㅏ') // false
- * canBeChosung('가') // false
- */
-export function canBeChosung(character: string): character is (typeof HANGUL_CHARACTERS_BY_FIRST_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_FIRST_INDEX, character);
-}
-
-/**
- * @name canBeChoseong
- * @description
- * 인자로 받은 문자가 초성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeChoseong(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeChoseong('ㄱ') // true
- * canBeChoseong('ㅃ') // true
- * canBeChoseong('ㄱㅅ') // false
- * canBeChoseong('ㅏ') // false
- * canBeChoseong('가') // false
- */
-export function canBeChoseong(character: string): character is (typeof HANGUL_CHARACTERS_BY_FIRST_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_FIRST_INDEX, character);
-}
-
-/**
- * @name canBeJungsung
- * @deprecated canBeJungseong을 사용해 주세요.
- * @description
- * 인자로 받은 문자가 중성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeJungsung(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeJungsung('ㅏ') // true
- * canBeJungsung('ㅗㅏ') // true
- * canBeJungsung('ㅏㅗ') // false
- * canBeJungsung('ㄱ') // false
- * canBeJungsung('ㄱㅅ') // false
- * canBeJungsung('가') // false
- */
-export function canBeJungsung(character: string): character is (typeof HANGUL_CHARACTERS_BY_MIDDLE_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_MIDDLE_INDEX, character);
-}
-
-/**
- * @name canBeJungseong
- * @description
- * 인자로 받은 문자가 중성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeJungseong(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeJungseong('ㅏ') // true
- * canBeJungseong('ㅗㅏ') // true
- * canBeJungseong('ㅏㅗ') // false
- * canBeJungseong('ㄱ') // false
- * canBeJungseong('ㄱㅅ') // false
- * canBeJungseong('가') // false
- */
-export function canBeJungseong(character: string): character is (typeof HANGUL_CHARACTERS_BY_MIDDLE_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_MIDDLE_INDEX, character);
-}
-
-/**
- * @name canBeJongsung
- * @deprecated canBeJongseong을 사용해 주세요.
- * @description
- * 인자로 받은 문자가 종성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeJongsung(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeJongsung('ㄱ') // true
- * canBeJongsung('ㄱㅅ') // true
- * canBeJongsung('ㅎㄹ') // false
- * canBeJongsung('가') // false
- * canBeJongsung('ㅏ') // false
- * canBeJongsung('ㅗㅏ') // false
- */
-export function canBeJongsung(character: string): character is (typeof HANGUL_CHARACTERS_BY_LAST_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_LAST_INDEX, character);
-}
-
-/**
- * @name canBeJongseong
- * @description
- * 인자로 받은 문자가 종성으로 위치할 수 있는 문자인지 검사합니다.
- * ```typescript
- * canBeJongseong(
- *   // 대상 문자
- *   character: string
- * ): boolean
- * ```
- * @example
- * canBeJongseong('ㄱ') // true
- * canBeJongseong('ㄱㅅ') // true
- * canBeJongseong('ㅎㄹ') // false
- * canBeJongseong('가') // false
- * canBeJongseong('ㅏ') // false
- * canBeJongseong('ㅗㅏ') // false
- */
-export function canBeJongseong(character: string): character is (typeof HANGUL_CHARACTERS_BY_LAST_INDEX)[number] {
-  return hasValueInReadOnlyStringList(HANGUL_CHARACTERS_BY_LAST_INDEX, character);
 }
 
 export function hasValueInReadOnlyStringList<T extends string>(list: readonly T[], value: string): value is T {
